Replace `hash_vec_merkle_tree()` with `hash_merkle_tree()`
